requirements.txt
asgiref==3.5.2
django==4.0.4
djangorestframework==3.13.1
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
# https://docs.djangoproject.com/en/4.1/ref/contrib/gis/install/geolibs/
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0